+++ include/info.h
@@ -21,10 +21,9 @@
 
 #include <stdio.h>
 
-#define VERSION  			"1.8.0"
+#define VERSION  			"2.0.0"
 #define AUTHOR  			"Alfredo Orozco de la Paz"
 #define AUTHOR_EMAIL		"alfredoopa@gmail.com"
-#define LAST_COMPILATION	"November-10-2018"
 
 #define RED				"\x1B[31m"
 #define BOLD_RED		"\x1B[1;31m"
@@ -47,8 +46,15 @@
  * Print the APOS versión on the command line
 */
 void show_version(){
-	printf(RESET GREEN "\n APOS" RESET " (AVR Project Open Source)"BOLD_CYAN" v%s\n" BOLD_RESET" Last compilation: " BOLD_CYAN "%s\n", VERSION, LAST_COMPILATION);
+	printf("APOS v%s (AVR Project Open Source)\n", VERSION);
+}
 
+/**
+ * Shows the APOS help in the command line
+*/
+void show_help() {
+	show_version();
+	printf("\n");
 	printf( BOLD_MAGENTA "      ___           ___           ___           ___     									\n");
 	printf("     /\\  \\         /\\  \\         /\\  \\         /\\  \\    						\n");
 	printf("    /::\\  \\       /::\\  \\       /::\\  \\       /::\\  \\   						\n");
@@ -64,28 +70,21 @@ void show_version(){
 	printf(BOLD_CYAN " Creates an AVR project template based on a makefile.\n Uses the avr-gcc toolchain and avrdude.\n\n");
 	printf(BOLD_RESET "   Usage:" BOLD_RED " apos " RESET "[" BOLD_YELLOW "OPTIONS" RESET "] <" BOLD_CYAN "Project_Name" RESET ">\n\n");
 	printf(BOLD_YELLOW"   -?"RESET","BOLD_YELLOW " -h" RESET " or" BOLD_YELLOW" ? "RESET"option shows help\n\n");
-}
-
-/**
- * Shows the APOS help in the command line
-*/
-void show_help() {
-	show_version();
 
 	printf(BOLD_YELLOW "--------------------------------------------\n\n" RESET);
 	printf(BOLD_YELLOW "   OPTIONS:\n\n" RESET);
-	printf(YELLOW "     -ap" RESET "      Define the ABOS Loader serial port (default: /dev/ttyUSB0)\n");
-	printf(YELLOW "     -ab" RESET "      Define the ABOS Loader baudrate (default: 38400)\n");
 	printf(YELLOW "     -b" RESET "       Create a blink project template\n");
+	printf(YELLOW "     -c" RESET "       Define the avrdude programmer (default: usbtiny)\n");
+	printf(YELLOW "     -e" RESET "       Define the Extended Fuse  (2 digit hex format)\n");
 	printf(YELLOW "     -f" RESET "       Define the CPU Frequency in Hz (default 16000000)\n");
 	printf(YELLOW "     -g" RESET "       Initialize a GIT repository in the project folder\n");
-	printf(YELLOW "     -ef" RESET "      Define the Extended Fuse  (2 digit hex format)\n");
-	printf(YELLOW "     -hf" RESET "      Define the High Fuse  (2 digit hex format)\n");
-	printf(YELLOW "     -lf" RESET "      Define the Low Fuse (2 digit hex format)\n");
+	printf(YELLOW "     -h" RESET "       Define the High Fuse  (2 digit hex format)\n");
+	printf(YELLOW "     -l" RESET "       Define the Low Fuse (2 digit hex format)\n");
 	printf(YELLOW "     -m" RESET "       Define the Microcontroller (default: atmega128)\n");
-	printf(YELLOW "     -p" RESET "       Define the avrdude programmer (default: usbtiny)\n");
+	printf(YELLOW "     -p" RESET "       Define the ABOS Loader serial port (default: /dev/ttyUSB0)\n");
+	printf(YELLOW "     -s" RESET "       Define the ABOS Loader baudrate (default: 38400)\n");
 	printf(YELLOW "     -v" RESET "       Show the APOS version\n");
-	printf(YELLOW "     -h,-?,? " RESET " Show this help\n\n");
+	printf(YELLOW "     -?,? " RESET " Show this help\n\n");
 	
 	printf(BOLD_CYAN "--------------------------------------------\n\n" RESET);
 	printf(BOLD_CYAN "   Makefile Options:\n\n" RESET);
@@ -104,13 +103,15 @@ void show_help() {
 	
 	printf(BOLD_MAGENTA "--------------------------------------------\n\n" RESET);
 	printf(BOLD_MAGENTA "   Examples: create a blink project\n\n" RESET);
-	printf("     Empty project:            " RED "apos " CYAN "blink\n" RESET);
+	printf("     Empty project:            " RED "apos " CYAN "example\n" RESET);
+	printf("     With git repository:      " RED "apos " YELLOW "-g " CYAN "example\n" RESET);
+	printf("     Using usbtiny programmer: " RED "apos " YELLOW "-c " RESET "usbtiny " CYAN "example\n" RESET);
 	printf("     Create a Blink project:   " RED "apos " YELLOW "-b " CYAN "blink\n" RESET);
-	printf("     Define MCU:               " RED "apos " YELLOW "-m " RESET "atmega32 " CYAN "blink\n" RESET);
-	printf("     Define CPU frequency:     " RED "apos " YELLOW "-m " RESET "atmega32 " YELLOW "-f " RESET "16000000 " CYAN "blink\n" RESET);
-	printf("     Define programmer:        " RED "apos " YELLOW "-m " RESET "atmega32 " YELLOW "-f " RESET "16000000 " YELLOW "-p " RESET "usbtiny " CYAN "blink\n" RESET);
-	printf("     Define fuses:             " RED "apos " YELLOW "-m " RESET "atmega32 " YELLOW "-f " RESET "16000000 " YELLOW "-p " RESET "usbtiny " YELLOW "-ef "RESET"ff "YELLOW"-lf "RESET"de "YELLOW"-hf "RESET"D9 "CYAN"blink\n" RESET);
-	printf("     Define ABOS params:       " RED "apos " YELLOW "-m " RESET "atmega32 " YELLOW "-f " RESET "16000000 " YELLOW "-ap "RESET"/dev/ttyACM0 "YELLOW "-ab " RESET "38400 "CYAN"blink\n" RESET);
+	printf("     Define MCU:               " RED "apos " YELLOW "-m " RESET "atmega32 " CYAN "example\n" RESET);
+	printf("     Define CPU frequency:     " RED "apos " YELLOW "-m " RESET "atmega32 " YELLOW "-f " RESET "16000000 " CYAN "example\n" RESET);
+	printf("     Define programmer:        " RED "apos " YELLOW "-m " RESET "atmega32 " YELLOW "-f " RESET "16000000 " YELLOW "-p " RESET "usbtiny " CYAN "example\n" RESET);
+	printf("     Define fuses:             " RED "apos " YELLOW "-m " RESET "atmega32 " YELLOW "-f " RESET "16000000 " YELLOW "-p " RESET "usbtiny " YELLOW "-e "RESET"ff "YELLOW"-l "RESET"de "YELLOW"-h "RESET"D9 "CYAN"example\n" RESET);
+	printf("     Define ABOS params:       " RED "apos " YELLOW "-m " RESET "atmega32 " YELLOW "-f " RESET "16000000 " YELLOW "-p "RESET"/dev/ttyACM0 "YELLOW "-s " RESET "38400 "CYAN"example\n" RESET);
 	printf("                \n");
 	printf(RESET" Created by " BOLD_CYAN "%s" MAGENTA " <" YELLOW "%s" MAGENTA ">\n\n" RESET, AUTHOR, AUTHOR_EMAIL);